Logging improvements (#120)
Completely refactors the logging system. Dev warnings are now implemented and the user can copy the detailed error/warning/critical message via a copy button to send the report. Source locations are now logged. Including line number and function signature.

using `atlas:logging` as the logging source now also creates a message to the user (except for `debug`. Which is silent and only appears during release builds upon the printing of a warning/error/critical message)

We have a lot of cleanup to do in order to make this not unbearable.